Move unique and exclude to slice.go
